place query bugfix, added generalSignup function
