+++ frontend/.env
@@ -1,4 +1,4 @@
 BROWSER=none
 REACT_EDITOR=none
 
-REACT_APP_GRAPHQL_API=http://localhost:3306/graphql
+REACT_APP_GRAPHQL_API=http://localhost:4000/graphql
